Lots of refactoring to simplify the data access layer code
Factored so that there is a single "main" file, which will connect to
the database via any supported ORM/DAL, simply by changing the imported
package. Removed the individual main files that were designed to execute
the data access code. Replaced with comments and commented package
imports in the single main file. Added a new interface to the common
package, which the ORM/DAL integrations will implement. Revamped both
the GORM and raw SQL packages to be in line with interface.